new branch ___ connect view <=> new parser

+++ src/model/kernel/parser.h
@@ -18,7 +18,7 @@ class Parser {
  public:
   void OpenFile(std::string &input_file);
   std::pair<size_t, size_t> PreParser(std::string &input_file);
-  std::pair<std::vector<double>, std::vector<size_t>> GetDataVetrtexAndFacet(std::string &input_file,
+  std::pair<std::vector<double>, std::vector<unsigned int>> GetDataVetrtexAndFacet(std::string &input_file,
                               size_t v_size,
                               size_t f_size);
   void WriteData();
